+++ pyclean/pyproject.toml
@@ -10,8 +10,13 @@ readme = "README.md"
 python = "^3.9"
 pandas = "^2.2.1"
 pytest = "^8.1.1"
+numpy = "^1.26.4"
 
 [tool.poetry.dev-dependencies]
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.1.1"
+pytest-cov = "^5.0.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
