--- pyclean/src/pyclean/changes.py
@@ -1,21 +1,0 @@
-import pandas as pd
-
-def compare_data(raw_data, processed_data):
-    # Compare number of missing values
-    raw_missing_values = raw_data.isnull().sum()
-    processed_missing_values = processed_data.isnull().sum()
-    missing_values_diff = processed_missing_values - raw_missing_values
-    
-    # Compare number of unique values
-    raw_unique_values = raw_data.nunique()
-    processed_unique_values = processed_data.nunique()
-    unique_values_diff = processed_unique_values - raw_unique_values
-    
-    # Combine results into a dictionary
-    comparison_results = {
-        'missing_values': missing_values_diff,
-        'unique_values': unique_values_diff
-        # Add more comparisons as needed
-    }
-    
-    return comparison_results
