+++ pyclean/tests/test_pyclean.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import numpy as np
 import pytest
 import sys
 sys.path.append('src/pyclean')
@@ -31,6 +32,28 @@ def test_remove_duplicates():
 
     # Check if duplicate rows are removed
     assert len(cleaned_df) == len(df.drop_duplicates())
+
+def generate_random_dataframe(num_rows, num_columns):
+    # Create a random DataFrame
+    data = {f'Column_{i}': np.random.randint(0, 100, size=num_rows) for i in range(num_columns)}
+    return pd.DataFrame(data)
+
+def test_compare_data():
+    # Generate random raw and processed dataframes for testing
+    num_rows = 100
+    raw_data = generate_random_dataframe(num_rows, 5)  
+    processed_data = generate_random_dataframe(num_rows, 5)  
+    
+    # Compare DataFrame
+    comparison_results = pyclean.compare_data(raw_data, processed_data)
+    
+    # Check if missing values comparison is correct
+    for column in raw_data.columns:
+        assert comparison_results['missing_values'][column] == processed_data[column].isnull().sum() - raw_data[column].isnull().sum()
+    
+    # Check if unique values comparison is correct
+    for column in raw_data.columns:
+        assert comparison_results['unique_values'][column] == processed_data[column].nunique() - raw_data[column].nunique()
 
 # Run tests
 if __name__ == "__main__":
